+++ src/main/kotlin/com/github/saundefined/bitrix_idea/dialogs/CreateTemplateDialogWrapper.kt
@@ -5,23 +5,22 @@ import com.github.saundefined.bitrix_idea.validation.TemplateCodeVerifier
 import com.intellij.ide.IdeView
 import com.intellij.ide.fileTemplates.FileTemplateManager
 import com.intellij.ide.fileTemplates.FileTemplateUtil
+import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.DialogWrapper
-import com.intellij.openapi.ui.ValidationInfo
 import com.intellij.psi.PsiDirectory
-import com.intellij.ui.components.JBLabel
-import com.intellij.ui.components.JBTextField
-import com.intellij.util.ui.JBUI
-import java.awt.GridBagConstraints
-import java.awt.GridBagLayout
+import com.intellij.ui.dsl.builder.AlignX
+import com.intellij.ui.dsl.builder.RowLayout
+import com.intellij.ui.dsl.builder.bindText
+import com.intellij.ui.dsl.builder.panel
+import com.intellij.ui.dsl.gridLayout.HorizontalAlign
 import java.util.*
 import javax.swing.JComponent
-import javax.swing.JPanel
 
 class CreateTemplateDialogWrapper : DialogWrapper(true) {
-    private var codeField = JBTextField(20)
-    private var nameField = JBTextField(20)
-    private var descriptionField = JBTextField(55)
+    var code: String = ""
+    var name: String = ""
+    var description: String = ""
 
     lateinit var view: IdeView
     lateinit var project: Project
@@ -33,64 +32,57 @@ class CreateTemplateDialogWrapper : DialogWrapper(true) {
     }
 
     override fun createCenterPanel(): JComponent {
-        val panel = JPanel(GridBagLayout())
-        val constraints = GridBagConstraints()
+        return panel {
+            row {
+                label(message("create.template.code"))
+                textField()
+                    .bindText(::code)
+                    .focused()
+                    .align(AlignX.FILL)
+                    .validationOnApply {
+                        val value = it.text
+                        when {
+                            TemplateCodeVerifier().verify(value)
+                                .not() -> error(message("create.template.code.validation.fail"))
 
-        constraints.anchor = GridBagConstraints.WEST
-        constraints.insets = JBUI.insets(10)
+                            else -> null
+                        }
+                    }
 
-        constraints.gridx = 0;
-        constraints.gridy = 0;
-        constraints.gridwidth = 1;
-        panel.add(JBLabel(message("create.template.code")), constraints);
+                label(message("create.template.name"))
+                textField()
+                    .bindText(::name)
+                    .align(AlignX.FILL)
+                    .validationOnApply {
+                        val value = it.text
+                        when {
+                            value === null -> error(message("create.template.name.validation.fail"))
+                            else -> null
+                        }
+                    }
+            }.layout(RowLayout.PARENT_GRID)
 
-        constraints.gridx = 1;
-        panel.add(codeField, constraints);
-
-        constraints.gridx = 2;
-        panel.add(JBLabel(message("create.template.name")), constraints);
-
-        constraints.gridx = 3;
-        panel.add(nameField, constraints);
-
-        constraints.gridx = 0;
-        constraints.gridy = 1;
-        panel.add(JBLabel(message("create.template.description")), constraints);
-
-        constraints.gridx = 1;
-        constraints.gridwidth = 3;
-        panel.add(descriptionField, constraints);
-
-        return panel
-    }
-
-    override fun doValidate(): ValidationInfo? {
-        if (TemplateCodeVerifier().verify(codeField.text).not()) {
-            return ValidationInfo(message("create.template.code.validation.fail"), codeField)
+            row(message("create.template.description")) {
+                textField()
+                    .bindText(::description)
+                    .align(AlignX.FILL)
+            }.layout(RowLayout.PARENT_GRID)
         }
-        if (nameField.text.isBlank()) {
-            return ValidationInfo(message("create.template.name.validation.fail"), nameField)
-        }
-
-        return null
     }
 
     override fun doOKAction() {
-        val validationInfo = doValidate()
-        if (validationInfo == null) {
-            val code = codeField.text
-            val name = nameField.text
-            val description = descriptionField.text
+        super.applyFields()
 
-            val templateManager = FileTemplateManager.getInstance(project)
+        val templateManager = FileTemplateManager.getInstance(project)
 
-            val properties: Properties = FileTemplateManager.getInstance(project).defaultProperties
-            properties.setProperty("TEMPLATE_NAME", name)
-            properties.setProperty("TEMPLATE_DESCRIPTION", description)
+        val properties: Properties = FileTemplateManager.getInstance(project).defaultProperties
+        properties.setProperty("TEMPLATE_NAME", name)
+        properties.setProperty("TEMPLATE_DESCRIPTION", description)
 
-            val templateDirectory = directory.createSubdirectory(code)
-
+        ApplicationManager.getApplication().runWriteAction {
             try {
+                val templateDirectory = directory.createSubdirectory(code)
+
                 val fileNames = listOf(
                     ".styles.php", "description.php", "footer.php", "header.php", "styles.css", "template_styles.css"
                 )
@@ -102,8 +94,9 @@ class CreateTemplateDialogWrapper : DialogWrapper(true) {
             } catch (e: Exception) {
                 e.printStackTrace()
             }
-
-            super.doOKAction()
         }
+
+        super.doOKAction()
     }
+
 }