+++ src/main/kotlin/com/github/saundefined/bitrix_idea/icons/BitrixIdeaIcons.kt
@@ -1,9 +1,0 @@
-package com.github.saundefined.bitrix_idea.icons
-
-import com.intellij.openapi.util.IconLoader.getIcon
-
-object BitrixIdeaIcons {
-    @JvmField
-    val Bitrix = getIcon("icons/bitrix.svg", javaClass)
-
-}