+++ src/main/kotlin/com/github/saundefined/bitrix_idea/validation/TemplateCodeVerifier.kt
@@ -2,6 +2,6 @@ package com.github.saundefined.bitrix_idea.validation
 
 class TemplateCodeVerifier {
     fun verify(input: String): Boolean {
-        return input.matches(Regex("^[a-zA-Z0-9_]+$")) && input.isNotBlank()
+        return input.matches(Regex("^[a-zA-Z0-9_-]+$")) && input.isNotBlank()
     }
 }