+++ src/main/kotlin/com/github/saundefined/bitrix_idea/validation/UrlVerifier.kt
@@ -2,6 +2,6 @@ package com.github.saundefined.bitrix_idea.validation
 
 class UrlVerifier {
     fun verify(input: String): Boolean {
-        return input.matches(Regex("^https?://\\S*\$")) || input.isEmpty()
+        return input.matches(Regex("^https?://\\S*\$"))
     }
 }