+++ CHANGELOG.md
@@ -2,16 +2,67 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Updated Bitrix CMS core
+- Update IntelliJ Platform 1.14.1
+- Update IntelliJ Platform 1.14.2
+- Update IntelliJ Platform 2.0.0
+- Update IntelliJ Platform 2.0.1
+
+## [2024.2.0]
+
+### Added
+
+- Creating Bitrix-related .gitignore file
+
+### Changed
+
+- Updated Bitrix CMS core
+- Update IntelliJ Platform 1.13.0
+- Update IntelliJ Platform 1.14.0
+- Support for PhpStorm 2024.2 EAP
+- Fix icons
+
+## [2024.1.0]
+
+### Changed
+
+- Updated Bitrix CMS core
+- Support for PhpStorm 2024.1 EAP 
+- Update IntelliJ Platform 1.12.0
+
+## [2023.7.0]
+
+### Changed
+
+- Updated Bitrix CMS core
+- Update IntelliJ Platform 1.11.3
+
+## [2023.6.0]
+
+### Changed
+
+- Update IntelliJ Platform 1.10.0
+- Update IntelliJ Platform 1.11.0
+- Update IntelliJ Platform 1.11.2
+- Updated Bitrix CMS core
+
+## [2023.5.0]
+
 ### Added
 
 - Module creating form
 - Template creating form
 - Simple component creating form
+- Add Bitrix Idea settings
+- Add localizations table to settings and create entity forms
 
 ### Changed
 
 - Add "Don't show again" action to startup notification
 - Update IntelliJ Platform 1.8.0
+- Update IntelliJ Platform 1.9.0
 
 ## [2023.4.0]